+++ Src/Private/Add-DiaHtmlLabel.ps1
@@ -20,32 +20,34 @@ function Add-DiaHTMLLabel {
         Author:         Jonathan Colon
         Twitter:        @jcolonfzenpr
         GitHub:         rebelinux
-    .PARAMETER Label
-        The string used to set the Diagram Title. This parameter is mandatory.
-    .PARAMETER SubgraphLabel
-        Switch to create a table used to add a logo to a Graphviz subgraph.
-    .PARAMETER IconType
-        Specifies the main diagram logo type.
-    .PARAMETER ImagesObj
-        Hashtable containing the IconName to IconPath translation.
-    .PARAMETER IconDebug
-        Enables the table debug mode if set to $true.
-    .PARAMETER IconPath
-        Optionally specifies the full path to the icon image file. If not provided, the default image path is used.
-    .PARAMETER IconWidth
-        Specifies the width of the subgraph icon. Default is 40.
-    .PARAMETER IconHeight
-        Specifies the height of the subgraph icon. Default is 40.
-    .PARAMETER Fontsize
-        Specifies the font size of the label. Default is 1.
-    .PARAMETER fontName
-        Specifies the font name for the cell text. Default is "Segoe Ui Black".
-    .PARAMETER fontColor
-        Specifies the font color for the cell text. Default is "#565656".
     .PARAMETER CellPadding
         The padding inside the table cell. Default is 5.
     .PARAMETER CellSpacing
         The spacing between table cells. Default is 5.
+    .PARAMETER Fontsize
+        Specifies the font size of the label. Default is 1.
+    .PARAMETER fontColor
+        Specifies the font color for the cell text. Default is "#565656".
+    .PARAMETER fontName
+        Specifies the font name for the cell text. Default is "Segoe Ui Black".
+    .PARAMETER IconDebug
+        Enables the table debug mode if set to $true.
+    .PARAMETER IconHeight
+        Specifies the height of the subgraph icon. Default is 40.
+    .PARAMETER IconPath
+        Optionally specifies the full path to the icon image file. If not provided, the default image path is used.
+    .PARAMETER IconType
+        Specifies the main diagram logo type.
+    .PARAMETER IconWidth
+        Specifies the width of the subgraph icon. Default is 40.
+    .PARAMETER ImageSizePercent
+        Set the image size in percent (100% is default). Requires IconPath when less than 100.
+    .PARAMETER ImagesObj
+        Hashtable containing the IconName to IconPath translation.
+    .PARAMETER Label
+        The string used to set the Diagram Title. This parameter is mandatory.
+    .PARAMETER SubgraphLabel
+        Switch to create a table used to add a logo to a Graphviz subgraph.
     #>
 
     [CmdletBinding()]
@@ -56,52 +58,61 @@ function Add-DiaHTMLLabel {
             HelpMessage = 'Please provide Label string to process'
         )]
         [string] $Label,
+
         [Parameter(
             Mandatory = $false,
             HelpMessage = 'Please provide Image hashtable to process'
         )]
         [Hashtable] $ImagesObj = @{},
+
         [Parameter(
             Mandatory = $false,
             HelpMessage = 'Please provide Icon type to process'
         )]
         [string] $IconType,
+
         [Parameter(
             Mandatory = $false,
-            HelpMessage = 'Please provide SubgraphLabel to process'
+            HelpMessage = 'Include the full path for the icon images (default is false)'
         )]
-        [Switch] $SubgraphLabel,
-        [Parameter(
-            Mandatory = $false,
-            HelpMessage = 'Allow to set a subgraph icon width'
-        )]
-        [int] $IconWidth,
+        [System.IO.FileInfo] $IconPath,
+
         [Parameter(
             Mandatory = $false,
             HelpMessage = 'Allow to set a subgraph icon height'
         )]
         [int] $IconHeight,
+
+        [Parameter(
+            Mandatory = $false,
+            HelpMessage = 'Allow to set a subgraph icon width'
+        )]
+        [int] $IconWidth,
+
         [Parameter(
             Mandatory = $false,
             HelpMessage = 'Enable the icon debug mode'
         )]
         [Alias("DraftMode")]
         [bool] $IconDebug,
+
         [Parameter(
             Mandatory = $false,
             HelpMessage = 'Allow to set a label font size'
         )]
         [int] $Fontsize = 14,
-        [Parameter(
-            Mandatory = $false,
-            HelpMessage = 'The cell text font name'
-        )]
-        [string] $fontName = "Segoe Ui Black",
+
         [Parameter(
             Mandatory = $false,
             HelpMessage = 'The cell text font color'
         )]
         [string] $fontColor = "#565656",
+
+        [Parameter(
+            Mandatory = $false,
+            HelpMessage = 'The cell text font name'
+        )]
+        [string] $fontName = "Segoe Ui Black",
 
         [Parameter(
             Mandatory = $false,
@@ -136,9 +147,15 @@ function Add-DiaHTMLLabel {
 
         [Parameter(
             Mandatory = $false,
-            HelpMessage = 'Include the full path for the icon images (default is false)'
+            HelpMessage = 'Please provide SubgraphLabel to process'
         )]
-        [System.IO.FileInfo] $IconPath
+        [Switch] $SubgraphLabel,
+
+        [Parameter(
+            Mandatory = $false,
+            HelpMessage = 'Set the image size in percent (100% is default)'
+        )]
+        [int] $TableBorder = 0
     )
 
     if ($IconType -eq 'NoIcon') {
@@ -156,35 +173,35 @@ function Add-DiaHTMLLabel {
 
     if (-not $SubgraphLabel) {
         if ($IconDebug) {
-            return "<TABLE border='0' cellborder='0' cellspacing='$($CellSpacing)' cellpadding='$($CellSpacing)'><TR><TD bgcolor='#FFCCCC' ALIGN='center' colspan='1'>Main Logo</TD></TR><TR><TD bgcolor='#FFCCCC' ALIGN='center' ><FONT FACE='$($fontName)' Color='$($fontColor)' POINT-SIZE='$($Fontsize)'>$Label</FONT></TD></TR><TR><TD ALIGN='center'><font color='red'>Debug ON</font></TD></TR></TABLE>"
+            return "<TABLE border='$TableBorder' cellborder='0' cellspacing='$($CellSpacing)' cellpadding='$($CellSpacing)'><TR><TD bgcolor='#FFCCCC' ALIGN='center' colspan='1'>Main Logo</TD></TR><TR><TD bgcolor='#FFCCCC' ALIGN='center' ><FONT FACE='$($fontName)' Color='$($fontColor)' POINT-SIZE='$($Fontsize)'>$Label</FONT></TD></TR><TR><TD ALIGN='center'><font color='red'>Debug ON</font></TD></TR></TABLE>"
         } elseif ($ICON -ne 'NoIcon') {
             if ($IconWidth -and $IconHeight) {
-                return "<TABLE border='0' cellborder='0' cellspacing='$($CellSpacing)' cellpadding='$($CellPadding)'><TR><TD ALIGN='center' colspan='1' fixedsize='true' width='$($IconWidth)' height='$($IconHeight)'><img src='$($ICON)'/></TD></TR><TR><TD ALIGN='center'><FONT FACE='$($fontName)' Color='$($fontColor)' POINT-SIZE='$($Fontsize)'>$Label</FONT></TD></TR></TABLE>"
+                return "<TABLE border='$TableBorder' cellborder='0' cellspacing='$($CellSpacing)' cellpadding='$($CellPadding)'><TR><TD ALIGN='center' colspan='1' fixedsize='true' width='$($IconWidth)' height='$($IconHeight)'><img src='$($ICON)'/></TD></TR><TR><TD ALIGN='center'><FONT FACE='$($fontName)' Color='$($fontColor)' POINT-SIZE='$($Fontsize)'>$Label</FONT></TD></TR></TABLE>"
 
             } elseif ($CalculatedImageSize) {
-                return "<TABLE border='0' cellborder='0' cellspacing='$($CellSpacing)' cellpadding='$($CellPadding)'><TR><TD ALIGN='center' colspan='1' fixedsize='true' width='$($CalculatedImageSize.Width)' height='$($CalculatedImageSize.Height)'><img src='$($ICON)'/></TD></TR><TR><TD ALIGN='center'><FONT FACE='$($fontName)' Color='$($fontColor)' POINT-SIZE='$($Fontsize)'>$Label</FONT></TD></TR></TABLE>"
+                return "<TABLE border='$TableBorder' cellborder='0' cellspacing='$($CellSpacing)' cellpadding='$($CellPadding)'><TR><TD ALIGN='center' colspan='1' fixedsize='true' width='$($CalculatedImageSize.Width)' height='$($CalculatedImageSize.Height)'><img src='$($ICON)'/></TD></TR><TR><TD ALIGN='center'><FONT FACE='$($fontName)' Color='$($fontColor)' POINT-SIZE='$($Fontsize)'>$Label</FONT></TD></TR></TABLE>"
 
             } else {
-                return "<TABLE border='0' cellborder='0' cellspacing='$($CellSpacing)' cellpadding='$($CellPadding)'><TR><TD ALIGN='center' colspan='1'><img src='$($ICON)'/></TD></TR><TR><TD ALIGN='center'><FONT FACE='$($fontName)' Color='$($fontColor)' POINT-SIZE='$($Fontsize)'>$Label</FONT></TD></TR></TABLE>"
+                return "<TABLE border='$TableBorder' cellborder='0' cellspacing='$($CellSpacing)' cellpadding='$($CellPadding)'><TR><TD ALIGN='center' colspan='1'><img src='$($ICON)'/></TD></TR><TR><TD ALIGN='center'><FONT FACE='$($fontName)' Color='$($fontColor)' POINT-SIZE='$($Fontsize)'>$Label</FONT></TD></TR></TABLE>"
 
             }
         } else {
-            return "<TABLE border='0' cellborder='0' cellspacing='$($CellSpacing)' cellpadding='$($CellSpacing)'><TR><TD ALIGN='center' colspan='1'>Main Logo</TD></TR><TR><TD ALIGN='center' ><FONT FACE='$($fontName)' Color='$($fontColor)' POINT-SIZE='$($Fontsize)'>$Label</FONT></TD></TR></TABLE>"
+            return "<TABLE border='$TableBorder' cellborder='0' cellspacing='$($CellSpacing)' cellpadding='$($CellSpacing)'><TR><TD ALIGN='center' ><FONT FACE='$($fontName)' Color='$($fontColor)' POINT-SIZE='$($Fontsize)'>$Label</FONT></TD></TR></TABLE>"
         }
     }
     if ($SubgraphLabel) {
         if ($IconDebug) {
-            return "<TABLE border='0' cellborder='0' cellspacing='$($CellSpacing)' cellpadding='$($CellSpacing)'><TR><TD bgcolor='#FFCCCC' ALIGN='center' colspan='1'>Subgraph Logo</TD><TD bgcolor='#FFCCCC' ALIGN='center'><FONT FACE='$($fontName)' Color='$($fontColor)' POINT-SIZE='$($Fontsize)'>$Label</FONT></TD></TR></TABLE>"
+            return "<TABLE border='$TableBorder' cellborder='0' cellspacing='$($CellSpacing)' cellpadding='$($CellSpacing)'><TR><TD bgcolor='#FFCCCC' ALIGN='center' colspan='1'>Subgraph Logo</TD><TD bgcolor='#FFCCCC' ALIGN='center'><FONT FACE='$($fontName)' Color='$($fontColor)' POINT-SIZE='$($Fontsize)'>$Label</FONT></TD></TR></TABLE>"
         } if ($ICON -ne 'NoIcon') {
             if ($IconWidth -and $IconHeight) {
-                return "<TABLE border='0' cellborder='0' cellspacing='$($SubgraphCellPadding)' cellpadding='$($SubgraphCellPadding)'><TR><TD ALIGN='center' colspan='1' fixedsize='true' width='$($IconWidth)' height='$($IconHeight)'><img src='$($ICON)'/></TD><TD ALIGN='center'><FONT FACE='$($fontName)' Color='$($fontColor)' POINT-SIZE='$($Fontsize)'>$Label</FONT></TD></TR></TABLE>"
+                return "<TABLE border='$TableBorder' cellborder='0' cellspacing='$($SubgraphCellPadding)' cellpadding='$($SubgraphCellPadding)'><TR><TD ALIGN='center' colspan='1' fixedsize='true' width='$($IconWidth)' height='$($IconHeight)'><img src='$($ICON)'/></TD><TD ALIGN='center'><FONT FACE='$($fontName)' Color='$($fontColor)' POINT-SIZE='$($Fontsize)'>$Label</FONT></TD></TR></TABLE>"
             } elseif ($CalculatedImageSize) {
-                return "<TABLE border='0' cellborder='0' cellspacing='$($SubgraphCellPadding)' cellpadding='$($SubgraphCellPadding)'><TR><TD ALIGN='center' colspan='1' fixedsize='true' width='$($CalculatedImageSize.Width)' height='$($CalculatedImageSize.Height)'><img src='$($ICON)'/></TD><TD ALIGN='center'><FONT FACE='$($fontName)' Color='$($fontColor)' POINT-SIZE='$($Fontsize)'>$Label</FONT></TD></TR></TABLE>"
+                return "<TABLE border='$TableBorder' cellborder='0' cellspacing='$($SubgraphCellPadding)' cellpadding='$($SubgraphCellPadding)'><TR><TD ALIGN='center' colspan='1' fixedsize='true' width='$($CalculatedImageSize.Width)' height='$($CalculatedImageSize.Height)'><img src='$($ICON)'/></TD><TD ALIGN='center'><FONT FACE='$($fontName)' Color='$($fontColor)' POINT-SIZE='$($Fontsize)'>$Label</FONT></TD></TR></TABLE>"
             } else {
-                return "<TABLE border='0' cellborder='0' cellspacing='$($SubgraphCellPadding)' cellpadding='$($SubgraphCellPadding)'><TR><TD ALIGN='center' colspan='1'><img src='$($ICON)'/></TD><TD ALIGN='center'><FONT FACE='$($fontName)' Color='$($fontColor)' POINT-SIZE='$($Fontsize)'>$Label</FONT></TD></TR></TABLE>"
+                return "<TABLE border='$TableBorder' cellborder='0' cellspacing='$($SubgraphCellPadding)' cellpadding='$($SubgraphCellPadding)'><TR><TD ALIGN='center' colspan='1'><img src='$($ICON)'/></TD><TD ALIGN='center'><FONT FACE='$($fontName)' Color='$($fontColor)' POINT-SIZE='$($Fontsize)'>$Label</FONT></TD></TR></TABLE>"
             }
         } else {
-            return "<TABLE border='0' cellborder='0' cellspacing='$($CellSpacing)' cellpadding='$($CellSpacing)'><TR><TD ALIGN='center' colspan='1'>Subgraph Logo</TD><TD ALIGN='center'><FONT FACE='$($fontName)' Color='$($fontColor)' POINT-SIZE='$($Fontsize)'>$Label</FONT></TD></TR></TABLE>"
+            return "<TABLE border='$TableBorder' cellborder='0' cellspacing='$($CellSpacing)' cellpadding='$($CellSpacing)'><TD ALIGN='center'><FONT FACE='$($fontName)' Color='$($fontColor)' POINT-SIZE='$($Fontsize)'>$Label</FONT></TD></TR></TABLE>"
         }
     }
 }